// front/services/models/api-documentation.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Global Testing
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ApiDocumentation
 */
export interface ApiDocumentation {
    /**
     * 
     * @type {string}
     * @memberof ApiDocumentation
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiDocumentation
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiDocumentation
     */
    'url'?: string;
}